Replace Linear with LazyLinear in fc1 layer
LazyLinear was the only viable solution I could find to the input dim problem on the first fully-connected layer in the Feature Extraction Module. At the first pass, LazyLinear(out_features) reads the shape of the input tensor and automatically matches the in_features to the last dim, unlike regular Linear, where the developer must match the dimension of in_features and the last dimension of the input tensor manually. After the first pass, it behaves like a regular Linear using the in_features it copied and out_features the developer passed as an argument.